# [Enhancement] Improve ORC column reader's performance and add benchmark