--- a/be/src/bench/README.md
+++ b/be/src/bench/README.md
@@ -1,6 +1,6 @@
 ### Usage
 ```
-sh build.sh --with-bench
+sh build.sh --be --with-bench
 
 find . -name 'runtime_filter_bench'
 ./build_Release/src/bench/output/runtime_filter_bench