--- a/be/src/formats/CMakeLists.txt
+++ b/be/src/formats/CMakeLists.txt
@@ -45,6 +45,8 @@ add_library(Formats STATIC
         orc/orc_min_max_decoder.cpp
         orc/utils.cpp
         orc/column_reader.cpp
+        orc/memory_stream/MemoryInputStream.cc
+        orc/memory_stream/MemoryOutputStream.cc
         parquet/column_chunk_reader.cpp
         parquet/column_converter.cpp
         parquet/column_reader.cpp