--- a/be/src/formats/orc/apache-orc/c++/src/Compression.cc
+++ b/be/src/formats/orc/apache-orc/c++/src/Compression.cc
@@ -35,6 +35,10 @@
 
 #include "Compression.hh"
 
+#ifdef __x86_64__
+#include <libdeflate.h>
+#endif
+
 #include <algorithm>
 #include <iomanip>
 #include <iostream>
@@ -42,6 +46,7 @@
 #include "Adaptor.hh"
 #include "LzoDecompressor.hh"
 #include "Utils.hh"
+#include "common/config.h"
 #include "lz4.h"
 #include "orc/Exceptions.hh"
 #include "wrap/snappy-wrapper.h"
@@ -723,6 +728,41 @@ void BlockDecompressionStream::NextDecompress(const void** data, int* size, size
     outputBufferLength = 0;
 }
 
+#ifdef __x86_64__
+class LibDeflateDecompressionStream : public BlockDecompressionStream {
+public:
+    LibDeflateDecompressionStream(std::unique_ptr<SeekableInputStream> inStream, size_t blockSize, MemoryPool& _pool,
+                                  ReaderMetrics* _metrics)
+            : BlockDecompressionStream(std::move(inStream), blockSize, _pool, _metrics) {
+        decompressor = libdeflate_alloc_decompressor();
+        if (decompressor == nullptr) {
+            throw std::logic_error("LibDeflate decompressor alloc failed");
+        }
+    }
+
+    ~LibDeflateDecompressionStream() override { libdeflate_free_decompressor(decompressor); }
+
+    std::string getName() const override {
+        std::ostringstream result;
+        result << "LibDeflate(" << getStreamName() << ")";
+        return result.str();
+    }
+
+protected:
+    uint64_t decompress(const char* inputPtr, uint64_t length, char* output, size_t maxOutputLength) override {
+        size_t actual = 0;
+        auto res = libdeflate_deflate_decompress(decompressor, inputPtr, length, output, maxOutputLength, &actual);
+        if (res != LIBDEFLATE_SUCCESS) {
+            throw ParseError("LibDefalte decompress failed");
+        }
+        return actual;
+    }
+
+private:
+    libdeflate_decompressor* decompressor;
+};
+#endif
+
 class SnappyDecompressionStream : public BlockDecompressionStream {
 public:
     SnappyDecompressionStream(std::unique_ptr<SeekableInputStream> inStream, size_t blockSize, MemoryPool& _pool,
@@ -1110,8 +1150,18 @@ std::unique_ptr<SeekableInputStream> createDecompressor(CompressionKind kind,
     case CompressionKind_NONE:
         return REDUNDANT_MOVE(input);
     case CompressionKind_ZLIB:
+#ifdef __x86_64__
+        if (starrocks::config::enable_orc_libdeflate_decompression) {
+            return std::unique_ptr<SeekableInputStream>(
+                    new LibDeflateDecompressionStream(std::move(input), blockSize, pool, metrics));
+        } else {
+            return std::unique_ptr<SeekableInputStream>(
+                    new ZlibDecompressionStream(std::move(input), blockSize, pool, metrics));
+        }
+#elif
         return std::unique_ptr<SeekableInputStream>(
                 new ZlibDecompressionStream(std::move(input), blockSize, pool, metrics));
+#endif
     case CompressionKind_SNAPPY:
         return std::unique_ptr<SeekableInputStream>(
                 new SnappyDecompressionStream(std::move(input), blockSize, pool, metrics));