… be/src/bench/CMakeLists.txt
… be/src/formats/orc/column_reader.cpp
… be/src/formats/orc/memory_stream/MemoryInputStream.cc
… be/src/formats/orc/memory_stream/MemoryInputStream.hh
… be/src/formats/orc/memory_stream/MemoryOutputStream.cc
… be/src/formats/orc/memory_stream/MemoryOutputStream.hh